+++ application/views/KrsDetail/detail.php
@@ -15,8 +15,10 @@
               <h6 class="card-text text-left"><b>Nama Jurusan : </b><?= $data['nama_jurusan']; ?></h6>
               <h6 class="card-text text-left"><b>Tahun Akademik : </b><?= $data['nama_tahun_akademik']; ?></h6>
               <h6 class="card-text text-left"><b>Status: </b><?= $data['status']; ?></h6>
-              <span style="float: right;">
-                <a href="<?= base_url(); ?>KrsDetail" class="btn btn-info"> Kembali </a>
+              <?php if ($this->session->userdata('id_role') != "2"): ?>
+                <span style="float: right;">
+                  <a href="<?= base_url(); ?>KrsDetail" class="btn btn-info"> Kembali </a>
+                <?php endif ?>
               </span>
             </div>
           </div>
@@ -24,50 +26,59 @@
       </div>
     </div>
   </center>
+</div>
+<div class="container">
   <div class="card shadow mb-4">
     <div class="card-header py-3">
+     <?php if ($this->session->userdata('id_role') != "2"): ?>
       <div class="input-group input-group-sm hidden-xs" style="width: 150px;">
         <a href="<?= base_url(); ?>KrsDetail/tambah/<?= encrypt_url($data['nim_mhs']); ?>" class=" btn btn-block btn-dark bg-info"><b>+ Data Baru</b></a>
       </div>
-    </div>
-    <div class="card shadow mb-4">
-      <div class="card-header py-3">
-        <div class="card-body">
-          <div class="table-responsive">
-            <table class="table table-bordered" width="100%" cellspacing="0">
-              <thead>
-                <tr>
-                  <th>No</th>
-                  <th>Nama Dosen</th>
-                  <th>Mata Kuliah</th>
-                  <th>Nilai SKS</th>
-                  <th>Predikat</th>
-                  <th>Status</th>
+    <?php endif ?>
+  </div>
+  <div class="card shadow mb-4">
+    <div class="card-header py-3">
+      <div class="card-body">
+        <div class="table-responsive">
+          <table class="table table-bordered" width="100%" cellspacing="0">
+            <thead>
+              <tr>
+                <th>No</th>
+                <th>Nama Dosen</th>
+                <th>Mata Kuliah</th>
+                <th>Nilai SKS</th>
+                <th>Predikat</th>
+                <th>Status</th>
+                <?php if ($this->session->userdata('id_role') != "2"): ?>
                   <th>Aksi</th>
-                </tr>
-              </thead>
-              <tbody>
-                <tr>
-                  <?php
-                  $no = 1;
-                  foreach ($nilai as $nl) : ?>
-                    <td><?= $no; ?></td>
-                    <td><?= $nl->nama_dosen; ?></td>
-                    <td><?= $nl->nama_matkul; ?></td>
-                    <td><?= $nl->nilai_krs; ?></td>
-                    <td><?= $nl->grade; ?></td>
-                    <td><?= $nl->status; ?></td>
+                <?php endif ?>
+              </tr>
+            </thead>
+            <tbody>
+              <tr>
+                <?php
+                $no = 1;
+                foreach ($nilai as $nl) : ?>
+                  <td><?= $no; ?></td>
+                  <td><?= $nl->nama_dosen; ?></td>
+                  <td><?= $nl->nama_matkul; ?></td>
+                  <td><?= $nl->nilai_krs; ?></td>
+                  <td><?= $nl->grade; ?></td>
+                  <td><?= $nl->status; ?></td>
+                  <?php if ($this->session->userdata('id_role') != "2"): ?>
                     <td>
                       <a href="<?= base_url(); ?>KrsDetail/ubah/<?= encrypt_url($nl->id_krs); ?>"><button type="button" class="btn btn-success btn-circle"><i class="fas fa-fw fa-check-circle"></i></button></a> <b>|</b>
                       <a href="<?= base_url(); ?>KrsDetail/hapus/<?= encrypt_url($nl->id_krs); ?>" class="tombol-hapus"><button type="button" class="btn btn-danger btn-circle tombol-hapus"><i class="fas fa-fw fa-trash tombol-hapus"></i></button></a>
                     </td>
+                  <?php endif ?>
                 </tr>
               </tbody>
               <?php $no++ ?>
             <?php endforeach; ?>
-            </table>
-          </div>
+          </table>
         </div>
       </div>
     </div>
   </div>
+</div>
+</div>
